requirements.txt
ruamel_yaml
git+https://github.com/globocom/argocd-client.git
python-keycloak
kubernetes
minio
numpy
omegaconf
pandas
pyhelm3
PyYAML
Requests
SQLAlchemy
toml
click
nltk
hydra-core
python-dotenv